+++ include/utils.h
@@ -6,5 +6,6 @@
 t_is_return	free_return(char **target, t_is_return boolean);
 double		ft_atod(char *str);
 int			count_elements(char **elements);
+char		**get_elements_by_type(t_rt_list *file, char *type);
 
 #endif
